refactor cac keystore utility in for reuse in other modules
