+++ cwms-aaa-client/src/test/java/mil/army/usace/hec/cwms/aaa/client/CacKeyManagerTest.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2022 Hydrologic Engineering Center
+ * Copyright (c) 2023 Hydrologic Engineering Center
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,10 +24,11 @@
 
 package mil.army.usace.hec.cwms.aaa.client;
 
+import mil.army.usace.hec.cwms.http.client.auth.CacKeyManager;
+import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
 
 final class CacKeyManagerTest {
