+++ cwms-aaa-client/src/test/java/mil/army/usace/hec/cwms/aaa/client/CwmsAAALogoutTest.java
@@ -29,6 +29,7 @@ import mil.army.usace.hec.cwms.http.client.ApiConnectionInfoBuilder;
 import mil.army.usace.hec.cwms.http.client.CookieJarFactory;
 import mil.army.usace.hec.cwms.http.client.MockHttpServer;
 import mil.army.usace.hec.cwms.http.client.SslSocketData;
+import mil.army.usace.hec.cwms.http.client.auth.CacKeyManagerUtil;
 import org.junit.jupiter.api.Test;
 
 import javax.net.ssl.KeyManager;
