+++ cwms-radar-client/src/test/java/mil/army/usace/hec/cwms/radar/client/controllers/TestController.java
@@ -24,20 +24,6 @@
 
 package mil.army.usace.hec.cwms.radar.client.controllers;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.security.KeyStore;
-import java.util.prefs.Preferences;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSocketFactory;
-import javax.net.ssl.TrustManagerFactory;
-import javax.net.ssl.X509TrustManager;
-import mil.army.usace.hec.cwms.aaa.client.CacKeyManagerUtil;
 import mil.army.usace.hec.cwms.aaa.client.CwmsAuthCookieCallback;
 import mil.army.usace.hec.cwms.aaa.client.CwmsLoginController;
 import mil.army.usace.hec.cwms.http.client.ApiConnectionInfo;
@@ -47,9 +33,24 @@ import mil.army.usace.hec.cwms.http.client.CookieJarFactory;
 import mil.army.usace.hec.cwms.http.client.MockHttpServer;
 import mil.army.usace.hec.cwms.http.client.PreferencesBackedCookieStore;
 import mil.army.usace.hec.cwms.http.client.SslSocketData;
+import mil.army.usace.hec.cwms.http.client.auth.CacKeyManagerUtil;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
+
+import javax.net.ssl.KeyManager;
+import javax.net.ssl.KeyManagerFactory;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManagerFactory;
+import javax.net.ssl.X509TrustManager;
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.security.KeyStore;
+import java.util.prefs.Preferences;
 
 abstract class TestController {
